requirements.txt
azure-identity
azure-mgmt-web
azure-mgmt-sql
azure-mgmt-storage
azure-mgmt-apimanagement